# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface NetworkStatus {
  blockNumber: number;
  gasPrice: string;
  chainId: number;
  networkName: string;
  rpcUrl: string;
  isConnected: boolean;
  peerCount: number;
}

export interface CompileRequest {
  /** Solidity source code */
  source: string;
  /** Contract name to extract from the compiled output */
  contractName: string;
}

export interface CompileResult {
  abi: unknown[];
  bytecode: string;
  contractName: string;
  warnings: string[];
}

export interface DeployRequest {
  abi: unknown[];
  bytecode: string;
  privateKey: string;
  constructorArgs?: unknown[];
}

export interface DeployResult {
  contractAddress: string;
  transactionHash: string;
  blockNumber: number;
  gasUsed: string;
}

export interface TransactionDetails {
  hash: string;
  from: string;
  to?: string;
  value: string;
  gasPrice: string;
  gasLimit: string;
  gasUsed?: string;
  blockNumber: number;
  blockHash?: string;
  status: string;
  timestamp?: number;
  input: string;
  nonce: number;
}

export interface AddressInfo {
  address: string;
  balanceWei: string;
  balanceSHM: string;
  transactionCount: number;
  isContract: boolean;
  code?: string;
}

export interface ContractCallRequest {
  contractAddress: string;
  abi: unknown[];
  functionName: string;
  args?: unknown[];
}

export interface ContractCallResult {
  /** The result of the call (can be any type) */
  result: unknown;
  functionName: string;
}

export interface ContractSendRequest {
  contractAddress: string;
  abi: unknown[];
  functionName: string;
  args?: unknown[];
  privateKey: string;
  value?: string;
}

export interface ContractSendResult {
  transactionHash: string;
  blockNumber?: number;
  gasUsed?: string;
  status: string;
}

export interface BlockInfo {
  number: number;
  hash: string;
  timestamp: number;
  transactionCount: number;
  gasUsed: string;
  gasLimit: string;
  miner?: string;
}

export interface RecentBlocksResult {
  blocks: BlockInfo[];
  latestBlock: number;
}