# lib/api-zod/src/generated/types/addressInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInfo {
  address: string;
  balanceWei: string;
  balanceSHM: string;
  transactionCount: number;
  isContract: boolean;
  code?: string;
}