# lib/api-zod/src/generated/types/blockInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockInfo {
  number: number;
  hash: string;
  timestamp: number;
  transactionCount: number;
  gasUsed: string;
  gasLimit: string;
  miner?: string;
}