# lib/api-zod/src/generated/types/compileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface CompileRequest {
  /** Solidity source code */
  source: string;
  /** Contract name to extract from the compiled output */
  contractName: string;
}