# lib/api-zod/src/generated/types/contractCallResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface ContractCallResult {
  /** The result of the call (can be any type) */
  result: unknown;
  functionName: string;
}