# lib/api-zod/src/generated/types/contractSendResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface ContractSendResult {
  transactionHash: string;
  blockNumber?: number;
  gasUsed?: string;
  status: string;
}