# lib/api-zod/src/generated/types/deployRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface DeployRequest {
  abi: unknown[];
  bytecode: string;
  privateKey: string;
  constructorArgs?: unknown[];
}