# lib/api-zod/src/generated/types/networkStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface NetworkStatus {
  blockNumber: number;
  gasPrice: string;
  chainId: number;
  networkName: string;
  rpcUrl: string;
  isConnected: boolean;
  peerCount: number;
}