# lib/api-zod/src/generated/types/recentBlocksResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */
import type { BlockInfo } from "./blockInfo";

export interface RecentBlocksResult {
  blocks: BlockInfo[];
  latestBlock: number;
}