# lib/api-zod/src/generated/types/transactionDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shardeum Developer Tools API
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionDetails {
  hash: string;
  from: string;
  to?: string;
  value: string;
  gasPrice: string;
  gasLimit: string;
  gasUsed?: string;
  blockNumber: number;
  blockHash?: string;
  status: string;
  timestamp?: number;
  input: string;
  nonce: number;
}